requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
prophet==1.1.5
statsmodels==0.14.0
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.0
python-dotenv==1.0.0
python-docx==0.8.11
